Move `PhysicalItemPF2e` to @item/physical/document (#3992)
Also:
- Move @item/runes to @item/physical/runes
- Move some types and values out of @item/armor/data to @item/armor/types and @item/armor/values
