+++ src/module/item/armor/data.ts
@@ -7,9 +7,8 @@ import {
     PhysicalSystemSource,
 } from "@item/physical/data";
 import { OneToFour, ZeroToThree } from "@module/data";
-import type { LocalizePF2e } from "@module/system/localize";
 import type { ArmorPF2e } from ".";
-import { OtherArmorTag } from "./types";
+import { ArmorCategory, ArmorGroup, ArmorTrait, BaseArmorType, OtherArmorTag, ResilientRuneType } from "./types";
 
 type ArmorSource = BasePhysicalItemSource<"armor", ArmorSystemSource>;
 
@@ -68,18 +67,10 @@ interface ArmorSystemData
     };
 }
 
-type ArmorTrait = keyof ConfigPF2e["PF2E"]["armorTraits"];
 interface ArmorTraitsSource extends PhysicalItemTraits<ArmorTrait> {
     otherTags?: OtherArmorTag[];
 }
 
 type ArmorTraits = Required<ArmorTraitsSource>;
 
-export type ArmorCategory = keyof ConfigPF2e["PF2E"]["armorTypes"];
-export type ArmorGroup = keyof ConfigPF2e["PF2E"]["armorGroups"];
-export type BaseArmorType = keyof typeof LocalizePF2e.translations.PF2E.Item.Armor.Base;
-export type ResilientRuneType = "" | "resilient" | "greaterResilient" | "majorResilient";
-
-const ARMOR_CATEGORIES = ["unarmored", "light", "medium", "heavy"] as const;
-
-export { ArmorData, ArmorSource, ArmorSystemData, ArmorSystemSource, ArmorTrait, ARMOR_CATEGORIES };
+export { ArmorData, ArmorSource, ArmorSystemData, ArmorSystemSource };
