+++ src/module/item/armor/values.ts
@@ -1,0 +1,3 @@
+const ARMOR_CATEGORIES = ["unarmored", "light", "medium", "heavy"] as const;
+
+export { ARMOR_CATEGORIES };
