+++ src/module/item/book/sheet.ts
@@ -1,5 +1,4 @@
-import { PhysicalItemSheetPF2e } from "@item/physical/sheet";
-import { PhysicalItemSheetData } from "@item/sheet/data-types";
+import { PhysicalItemSheetData, PhysicalItemSheetPF2e } from "@item/physical/sheet";
 import { BookPF2e } from "./document";
 
 export class BookSheetPF2e extends PhysicalItemSheetPF2e<BookPF2e> {
