+++ src/module/item/consumable/sheet.ts
@@ -1,6 +1,5 @@
 import { ConsumablePF2e } from "@item";
-import { PhysicalItemSheetPF2e } from "@item/physical/sheet";
-import { PhysicalItemSheetData } from "@item/sheet/data-types";
+import { PhysicalItemSheetData, PhysicalItemSheetPF2e } from "@item/physical";
 
 export class ConsumableSheetPF2e extends PhysicalItemSheetPF2e<ConsumablePF2e> {
     override async getData(options?: Partial<DocumentSheetOptions>): Promise<ConsumableSheetData> {
