+++ src/module/item/container/sheet.ts
@@ -1,5 +1,4 @@
-import { PhysicalItemSheetPF2e } from "@item/physical/sheet";
-import { PhysicalItemSheetData } from "@item/sheet/data-types";
+import { PhysicalItemSheetData, PhysicalItemSheetPF2e } from "@item/physical";
 import { ContainerPF2e } from ".";
 
 export class ContainerSheetPF2e extends PhysicalItemSheetPF2e<ContainerPF2e> {
