+++ src/module/item/sheet/data-types.ts
@@ -1,10 +1,4 @@
-import { ItemPF2e, PhysicalItemPF2e } from "@item";
-import { AncestryPF2e } from "@item/ancestry";
-import { FeatPF2e } from "@item/feat";
-import { HeritagePF2e } from "@item/heritage";
-import { ItemActivation } from "@item/physical/data";
-import { MaterialGradeData } from "@item/physical/materials";
-import { PreciousMaterialGrade } from "@item/physical/types";
+import { AncestryPF2e, FeatPF2e, HeritagePF2e, ItemPF2e } from "@item";
 import { Rarity } from "@module/data";
 import { RuleElementSource } from "@module/rules";
 import { SheetOptions } from "@module/sheet/helpers";
@@ -52,21 +46,6 @@ export interface ItemSheetDataPF2e<TItem extends ItemPF2e> extends ItemSheetData
     proficiencies: ConfigPF2e["PF2E"]["proficiencyLevels"];
 }
 
-export interface PhysicalItemSheetData<TItem extends PhysicalItemPF2e> extends ItemSheetDataPF2e<TItem> {
-    isPhysical: true;
-    basePriceString: string;
-    priceString: string;
-    actionTypes: ConfigPF2e["PF2E"]["actionTypes"];
-    actionsNumber: ConfigPF2e["PF2E"]["actionsNumber"];
-    bulkTypes: ConfigPF2e["PF2E"]["bulkTypes"];
-    frequencies: ConfigPF2e["PF2E"]["frequencies"];
-    sizes: ConfigPF2e["PF2E"]["actorSizes"];
-    stackGroups: ConfigPF2e["PF2E"]["stackGroups"];
-    usage: ConfigPF2e["PF2E"]["usageTraits"];
-    bulkDisabled: boolean;
-    activations: { action: ItemActivation; id: string; base: string }[];
-}
-
 export interface FeatSheetData extends ItemSheetDataPF2e<FeatPF2e> {
     featTypes: ConfigPF2e["PF2E"]["featTypes"];
     actionTypes: ConfigPF2e["PF2E"]["actionTypes"];
@@ -84,13 +63,3 @@ export interface HeritageSheetData extends ItemSheetDataPF2e<HeritagePF2e> {
     ancestry: AncestryPF2e | null;
     ancestryRefBroken: boolean;
 }
-
-type MaterialSheetEntry = {
-    label: string;
-    grades: Partial<Record<PreciousMaterialGrade, MaterialGradeData>>;
-};
-
-export type MaterialSheetData = {
-    value: string;
-    materials: Record<string, MaterialSheetEntry>;
-};
