+++ src/module/item/treasure/sheet.ts
@@ -1,6 +1,5 @@
-import { TreasurePF2e } from "@item";
-import { PhysicalItemSheetPF2e } from "@item/physical/sheet";
-import { PhysicalItemSheetData } from "@item/sheet/data-types";
+import { PhysicalItemSheetData, PhysicalItemSheetPF2e } from "@item/physical";
+import { TreasurePF2e } from ".";
 
 export class TreasureSheetPF2e extends PhysicalItemSheetPF2e<TreasurePF2e> {
     override async getData(options?: Partial<DocumentSheetOptions>): Promise<TreasureSheetData> {
