+++ src/module/item/weapon/sheet.ts
@@ -1,13 +1,15 @@
-import { WEAPON_MATERIAL_VALUATION_DATA } from "@item/physical/materials";
-import { PhysicalItemSheetPF2e } from "@item/physical/sheet";
-import { PhysicalItemSheetData } from "@item/sheet/data-types";
-import { CoinsPF2e } from "@item/physical/helpers";
+import {
+    CoinsPF2e,
+    PhysicalItemSheetData,
+    PhysicalItemSheetPF2e,
+    WEAPON_MATERIAL_VALUATION_DATA,
+} from "@item/physical";
 import { OneToFour, OneToThree } from "@module/data";
-import { setHasElement } from "@util";
+import { createSheetTags, SheetOptions } from "@module/sheet/helpers";
 import { LocalizePF2e } from "@system/localize";
+import { setHasElement } from "@util";
 import { WeaponPF2e } from ".";
 import { WeaponPropertyRuneSlot } from "./data";
-import { createSheetTags, SheetOptions } from "@module/sheet/helpers";
 import { MANDATORY_RANGED_GROUPS, WEAPON_RANGES } from "./values";
 
 export class WeaponSheetPF2e extends PhysicalItemSheetPF2e<WeaponPF2e> {
